Add files via upload
This code allows you to enter two fits files and subtract them. User needs to define a center for each image around which the difference image will be calibrated. The final image can be analyzed in ds9 for variable stars. Methodology for analysis can be found in the Methods section of "Creating a Pipeline to Search for Dwarf Nova Eruptions in Globular Cluster M15," also in this repository.