Add files via upload
This code allows the user to compare a list of stellar objects with a list of known objects. The intention is to identify which sources from the original list are undiscovered. This is written for a case in which the comparison dataset is a list of known rr lyra. Repeat sources will be removed, resulting in a final table of unique objects.